add event name shortcode
